Move button bar items to image gallery header
